+++ backend/substrapp/views/algo.py
@@ -23,11 +23,11 @@ from substrapp.models import Algo
 from substrapp.orchestrator import get_orchestrator_client
 from substrapp.serializers import AlgoSerializer
 from substrapp.utils import get_hash
+from substrapp.views.filters_utils import CharInFilter
+from substrapp.views.filters_utils import ChoiceInFilter
+from substrapp.views.filters_utils import MatchFilter
 from substrapp.views.filters_utils import ProcessPermissionFilter
 from substrapp.views.utils import ApiResponse
-from substrapp.views.utils import CharInFilter
-from substrapp.views.utils import ChoiceInFilter
-from substrapp.views.utils import MatchFilter
 from substrapp.views.utils import PermissionMixin
 from substrapp.views.utils import ValidationExceptionError
 from substrapp.views.utils import get_channel_name