+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Filtering on compute tasks metadata (#1131)
+
 ### Fixed
 
 - build psycopg2 from source (make libpq-dev package available in docker image) (#1132)